Implement new color picker #24
New color picket is a tool button on the main tool bar that
provides a popup menu with a color grid and size slider, depending
on the selected tool. Some of the tools get also a checkbox for
fill and a more button to add more colors to the color grid.