Fixed removing painter items
Sometimes when removing painter items they were removed from the
scene but were still shown on the view, eventhough the items had
no parents. Deleting would have resolved the issue but we need the
items for the undo/redo framework. Solution was to hide the items
when removing them from the scene and showing them again when
adding them to the scene.