Added wayland support for KDE and Gnome #56
Under Wayland taking screenshots is much more complicated as
wayland is more secure and doen't allow taking screenshots
directly. In order to take screenshots, we rely on the
compositor to provide the screenshot, which we query via DBus
calls. The screenshot support under Gnome Wayland is more advanced
as the KDE support so we support there more capture options.
Both implementation, KDE and Gnome, need more testing but the
current version  should be enough for the beginning.