Make UNIX AND NOT APPLE define a UNIX_X11
For most intents and purposes, UNIX AND NOT APPLE means "Linux",
but that doesn't cover all of them: the BSDs and Illumos have
the same X11 and Wayland and GNOME and KDE and OpenDesktop libraries
as Linux does. All the code that #ifdefs on __linux__ really means
"UNIX AND NOT APPLE", so give it a (not-amazingly-well-named)
define to hang on to instead.